ctx/chore(ci): add a basic testing/validation framework (#76)
Adds a basic testing and validation framework.  As I was looking through the helm testing framework I realized that I wanted something a little more flexible so I have extended the generation tools that we use internally.  With this new approach you may specify different value combinations when required to ensure that the templates are rendered as expected.  The naming convention for the test values files are `<chart>.<test_name>.yaml`.

When a test is created the expected file can be generated and manually verified by using `make generate-expected`.