feat(serving): add serving CRDs, optionally install knative-operator (#83)
Adds values to control which CRDs are created.  My thought process here is that there may be some value in controlling whether or not this chart creates the `FlyteWorkflow` CRD, for example to support migrations from Flyte to Union.

The real driver for this change is to optionally create the `knative` serving CRDs.  I don't love just importing this from a Kustomization manifest, but as far as I can tell the `knative` project doesn't maintain a Helm chart that we could put a dependency on.

Also optionally installs `knative-operator` with the `dataplane` chart.