chore(serving): Move knative-operator to its own chart (#88)
We want to avoid having the `dataplane` chart create lots of namespaces, and ideally we'd run `knative-operator` in a dedicated namespace.  This change introduces a new chart, which is just a wrapper around the `knative-operator` Kustomize manifests.

Users can opt to manage it themselves, or use our provided chart.  For now, we'll leave the CRDs bundled in `dataplane-crds` to avoid yet another CRD-specific chart.